requirements.txt
# AutoGenesis - Unified Requirements
# Core dependencies for all platforms

# MCP Server
mcp==1.23.0
anyio>=4.9.0

# Testing Framework
behave==1.2.6
selenium>=4.0.0

# Mobile/Mac Testing (Appium)
appium-python-client>=2.0.0
browserstack-sdk==1.23.8

# Windows Testing
pywinauto>=0.6.9; sys_platform == 'win32'
pyautogui>=0.9.54

# LLM Integration
langchain>=0.1.20
langchain-openai>=0.0.8
langchain-anthropic>=0.0.8
langchain-google-genai>=0.0.8
anthropic>=0.22.0
openai>=1.0.0
google-generativeai>=0.4.0

# Utilities
psutil
pillow>=11.1.0
janus>=0.6.0

# Memory profiling (optional)
Pympler==1.1
